[Fix] fix quantity, data type of MerchandiseReturnProposal

+++ api/Entities/MerchandiseReturnDetail.cs
@@ -3,8 +3,8 @@ namespace api.Entities
     public class MerchandiseReturnDetail : BaseEntity
     {
         public int Id { get; set; }
-        public int Quantity { get; set; }
-        public int QuantityReturned { get; set; }
+        public double Quantity { get; set; }
+        public double QuantityReturned { get; set; }
         public string Description { get; set; }
         public int MerchandiseReturnProposalId { get; set; }
         public MerchandiseReturnProposal MerchandiseReturnProposal { get; set; }
